+++ frontend/src/classes/Data.js
@@ -76,11 +76,4 @@ export default class Data {
         'watchlist-admin',
         'watchlist-manager',
     ]
-
-    static navigationParent = {
-        root: 'root',
-        management: 'management',
-        administration: 'administration',
-        member_data: 'member_data',
-    }
 }